file-processor/requirements.txt
python-dotenv
pika
Pillow